pyproject.toml
[project]
name = "mainframe-knowledgebase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.8"
dependencies = [
    "fastapi>=0.121.0",
    "langchain>=1.0.3",
    "langchain-core>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-openai>=1.0.1",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.0.0",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uvicorn[standard]>=0.38.0",
]
